feat: bump `fuel-core` to `0.42.0` (#1628)
# Release notes

In this release, we:

- Bumped `fuel-core` to `0.42.0`
- Added support for simulating calls at specific block height

# Breaking Changes

- **Error Variant Renaming:**
- The transaction `Reason` error variant previously named `Reverted` in
transaction errors is now renamed to `Failure`.
- The associated `revert_id` field is now an `Option<u64>` rather than a
plain `u64`.
- **Usage:** Update pattern matches and error handling to check for
`Error::Transaction(Reason::Failure { .. })` and handle the optional
`revert_id` accordingly.

- **Execution Simulation API Changes:**
- The simulation execution mode has changed from enum variants to a
struct with helper functions:
- Replace **`Execution::Realistic`** with **`Execution::realistic()`**.
- Replace **`Execution::StateReadOnly`** with
**`Execution::state_read_only()`**.
- To simulate at a specific block height, chain the call with
`.at_height(height)`. Note that the node needs to support historical
execution.
- The `dry_run_opt` method's `utxo_validation` argument changed to
`Option<bool>` and a new argument `at_height: Option<BlockHeight>` is
added.
- The method **`determine_missing_contracts`** no longer accepts an
optional parameter; it is now called without arguments and will
determine all missing contracts in one simulation.

- **`tx_status` changes**
    - The `Revert` variant has been replaced by the `Failure` variant.
- New variants have been added: `PreconfirmationSuccess` and
`PreconfirmationFailure`.
    - The `revert_id` type changed to `Option<u64>`
  - **Usage:** Update your pattern matching and error handling logic.

- **Amount Type Updates:**
- Functions across accounts, wallets, predicates, and providers have
switched from using `u64` to `u128` for asset amounts.
- **Usage:** Modify function calls (e.g., `get_asset_inputs_for_amount`,
`adjust_for_fee`, etc.) to pass amounts as `u128` (or convert them using
`.into()` if needed).

- **Node Configuration Adjustments:**
- The `NodeConfig` struct now includes a new boolean field:
`historical_execution`.
- **Usage:** To use `historical_execution` launch a node (with `rocksdb`
enabled) and the `historical_execution` flag set.

+++ packages/fuels-accounts/src/wallet.rs
@@ -88,7 +88,7 @@ mod unlocked {
         async fn get_asset_inputs_for_amount(
             &self,
             asset_id: AssetId,
-            amount: u64,
+            amount: u128,
             excluded_coins: Option<Vec<CoinTypeId>>,
         ) -> Result<Vec<Input>> {
             Ok(self
@@ -157,7 +157,7 @@ mod locked {
         async fn get_asset_inputs_for_amount(
             &self,
             asset_id: AssetId,
-            amount: u64,
+            amount: u128,
             excluded_coins: Option<Vec<CoinTypeId>>,
         ) -> Result<Vec<Input>> {
             Ok(self
